fixed assembly multiply routine
the routine assumed row-major matrices, but I'm using column-major ones. This led to a confusing interface, where the arguments had to be supplied in the reverse order that they were being multiplied in. Also made it so that the destination is not overwritten until the multiplication is complete, meaning that the destination can now be either source with no problems